+++ tests/manifest/test-manifest.ts
@@ -6,7 +6,7 @@ import { integrationTests } from './fragments/integration';
 import { apiTests } from './fragments/api';
 import { e2eTests } from './fragments/e2e';
 import { androidTests } from './fragments/android';
-import { vendor_deviceTests as vendorDeviceTests } from './fragments/vendor-device';
+import { vendorDeviceTests } from './fragments/vendor-device';
 import { postreleaseTests } from './fragments/postrelease';
 
 export const testManifest: TestEntry[] = [